feat(tests): added full test modeling and changed file naming convention
