+++ test/fixtures/subscription-field-rule-config-test.json
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "validateSubscriptionType": false,
+    "validateSubscriptionFields": true
+  }
+}
