+++ test/fixtures/subscription-field-schema-test.graphql
@@ -1,0 +1,12 @@
+# This is a Subscription documentation
+type Subscription {
+  snackAdded: Snack!
+  snackUpdated: Snack!
+  snackDeleted: ID!
+}
+
+schema {
+  query: Query
+  mutation: Mutation
+  subscription: Subscription
+}
