+++ test/fixtures/subscription-type-rule-config-test.json
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "validateSubscriptionType": true,
+    "validateSubscriptionFields": false
+  }
+}
